requirements.txt
black
discord.py
ffmpeg-python
google-api-python-client
isodate
PyNaCl
python-dotenv
pytubefix
yt-dlp
